fixes exp10

+++ packages/secret_math/src/math/core.rs
@@ -444,7 +444,7 @@ pub fn muldiv(x: U256, y: U256, mut denominator: U256) -> StdResult<U256> {
     /// @param x - integer between 0 and 32
     ///
     /// @return result 10^x as U256
-    pub const fn exp10(x: u32) -> U256 {
+    pub const fn exp10(x: u8) -> U256 {
             match x {
                 0 => QUINTILLIONTH,
                 1 => HUN_QUADTH,
